pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name="gemini-rotate"
version="0.1.0"
description = "A lightweight Python library for Google Gemini API key rotation and model fallback."
readme = "README.md"
license = "MIT"
authors = [
  { name = "Jayed", email = "jayedbinjahangir@gmail.com" }
]
dependencies = ["google-genai", "python-dotenv"]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/jayeeed/gemini-rotate"

[tool.setuptools.packages.find]
where = ["."]
